requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic>=2.7.4
pydantic-settings==2.1.0

# LangGraph and LangChain (compatible versions)
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Logging
structlog==24.1.0

# YAML parsing
pyyaml==6.0.1

# Async scheduling for reminders
apscheduler==3.10.4

# Redis (optional, for production state management)
redis==5.0.1
bcrypt>=4.0.0
